requirements.txt
# Core dependencies
unstructured[pdf]>=0.10.0
python-docx>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10

# CLI & UI
typer>=0.9.0
rich>=13.0.0
streamlit>=1.28.0

# Config & validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM & RAG
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Document processing
openpyxl>=3.1.0
Pillow>=10.0.0
tabula-py>=2.8.0

# Dev dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0




